Resolve "audit/delete orphaned resources that exist in nsx"

See merge request ukfast/api.ukfast/ecloud!1387

+++ app/Console/Commands/Nsx/UndeployOrphanedResources.php
@@ -48,6 +48,7 @@ class UndeployOrphanedResources extends Command
             });
 
         if ($networks->count() > 0) {
+            $this->markedForDeletion = [];
             $this->processDeletion($networks, 'router');
         }
 
@@ -168,17 +169,19 @@ class UndeployOrphanedResources extends Command
                         if (!$this->option('test-run')) {
                             $resource = $resources->get($id);
 
-                            $endpoint = match ($resource::class) {
-                                Router::class => 'policy/api/v1/infra/tier-1s/' . $resource->id,
-                                Network::class => 'policy/api/v1/infra/tier-1s/' . $resource?->router?->id . '/segments/' . $resource->id,
-                            };
+                            if ($resource) {
+                                $endpoint = match ($resource::class) {
+                                    Router::class => 'policy/api/v1/infra/tier-1s/' . $resource->id,
+                                    Network::class => 'policy/api/v1/infra/tier-1s/' . $resource?->router?->id . '/segments/' . $resource->id,
+                                };
 
-                            try {
-                                $resource->availabilityZone->nsxService()->delete($endpoint);
-                            } catch (\Exception $e) {
-                                if ($e->hasResponse() && $e->getResponse()->getStatusCode() != 404) {
-                                    $reason = null;
-                                    $failedDeletes[] = [$resource->id, $resource->name, $e->getMessage()];
+                                try {
+                                    $resource->availabilityZone->nsxService()->delete($endpoint);
+                                } catch (\Exception $e) {
+                                    if ($e->hasResponse() && $e->getResponse()->getStatusCode() != 404) {
+                                        $reason = null;
+                                        $failedDeletes[] = [$resource->id, $resource->name, $e->getMessage()];
+                                    }
                                 }
                             }
                         }
